+++ packages/terra-core-docs/src/terra-dev-site/doc/toggle/example/AnimatedToggle.jsx
@@ -1,17 +1,24 @@
 import React, { useState } from 'react';
-import IconInformation from 'terra-icon/lib/icon/IconInformation';
 import Toggle from 'terra-toggle';
+import IconInformation from 'terra-icon/lib/icon/IconInformation';
+import { KEY_TAB } from 'keycode-js';
+import './ToggleExample.module.scss';
 
-const ToggleDefault = () => {
+const AnimatedToggle = () => {
   const [isOpen, setIsOpen] = useState(false);
-
+  const [focused, setFocus] = useState(true);
   const handleOnClick = () => {
     setIsOpen(!isOpen);
+  };
+  const handleOnKeyDown = (event) => {
+    if (event.nativeEvent.keyCode === KEY_TAB) {
+      setFocus(true);
+    }
   };
 
   return (
     <div>
-      <IconInformation onClick={handleOnClick} aria-expanded={isOpen} aria-controls="toggle" />
+      <IconInformation data-show-focus-styles={focused} focusable a11yLabel="information Icon" tabIndex="0" role="button" onKeyDown={handleOnKeyDown} onMouseDown={() => { setFocus(false); }} onClick={handleOnClick} aria-expanded={isOpen} aria-controls="toggle" />
       {/**
       * The aria-expanded state is used on the triggering component to indicate the contents are collapsible, and whether a region is currently expanded or collapsed
       */}
@@ -27,4 +34,4 @@ const ToggleDefault = () => {
   );
 };
 
-export default ToggleDefault;
+export default AnimatedToggle;